server/backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
pydantic-settings==2.7.1
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.1
redis[hiredis]==5.2.1
celery[redis]==5.4.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.1
python-multipart==0.0.20
httpx==0.28.1
orjson==3.10.13
email-validator==2.2.0
slowapi==0.1.9
structlog==24.4.0
python-json-logger==3.2.1
pytest==8.3.4
pytest-asyncio==0.25.0
pytest-cov==6.0.0
factory-boy==3.3.1
